requirements.txt
-e .
elementpath==4.1.5
h5py==3.10.0
memory-profiler==0.61.0
PyYAML==6.0.1
xmlschema==2.5.0
pytest